requirements.txt
pandas
numpy
flask
scikit-learn
seaborn
datetime
mysql_connector
sqlalchemy
-e .